+++ packages/client/src/util/toDataValue.ts
@@ -29,6 +29,7 @@ import {
   isMediaUpload,
 } from "../object/mediaUpload.js";
 import { getWireObjectSet, isObjectSet } from "../objectSet/createObjectSet.js";
+import { isScenarioClient } from "../scenarios/ScenarioClient.js";
 import { isInterfaceActionParam } from "./interfaceUtils.js";
 import { isObjectSpecifiersObject } from "./isObjectSpecifiersObject.js";
 import { isOntologyObjectV2 } from "./isOntologyObjectV2.js";
@@ -147,6 +148,10 @@ export async function toDataValue(
       objectTypeApiName: value.$objectType,
       primaryKeyValue: value.$primaryKey,
     };
+  }
+
+  if (isScenarioClient(value)) {
+    return value.getScenarioReference();
   }
 
   // TODO (during queries implementation)