+++ packages/react-components-storybook/src/stories/ActionForm/ActionForm.stories.tsx
@@ -384,8 +384,9 @@ function UnsupportedFieldsActionFormStory(): React.ReactElement {
 }
 
 const meta = {
-  title: "Beta/ActionForm/Usage",
+  title: "Components/ActionForm/Usage",
   component: UpdateEmployeeActionFormStory,
+  tags: ["beta"],
   parameters: {
     controls: {
       expanded: true,