+++ packages/react-components-storybook/src/stories/ActionForm/BaseForm.stories.tsx
@@ -198,8 +198,9 @@ function BaseFormSubmissionOutput(): React.ReactElement {
 }
 
 const meta: Meta<BaseFormStoryProps> = {
-  title: "Beta/ActionForm/BaseForm",
+  title: "Components/ActionForm/BaseForm",
   component: BaseForm,
+  tags: ["beta"],
   decorators: [
     (Story) => (
       <FormStoryLayout output={<BaseFormSubmissionOutput />}>