+++ packages/react-components-storybook/src/stories/ImageViewer/ImageViewer.stories.tsx
@@ -76,8 +76,9 @@ function createMockImageMedia(
 }
 
 const meta: Meta<ImageViewerMediaProps> = {
-  title: "Beta/DocumentViewer/Renderers/ImageViewer",
+  title: "Components/DocumentViewer/Renderers/ImageViewer",
   component: ImageViewer,
+  tags: ["beta"],
   parameters: {
     controls: { expanded: true },
   },