+++ packages/react-components-storybook/src/stories/ObjectTable/ObjectTable.stories.tsx
@@ -45,8 +45,9 @@ import {
 type EmployeeTableProps = ObjectTableProps<typeof Employee>;
 
 const meta: Meta<EmployeeTableProps> = {
-  title: "Beta/ObjectTable/Features",
+  title: "Components/ObjectTable/Features",
   component: ObjectTable,
+  tags: ["beta"],
   parameters: {
     msw: {
       handlers: [...fauxFoundry.handlers],