+++ packages/create-app.template.react/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @osdk/create-app.template.react
 
+## 2.25.0
+
+## 2.24.0
+
 ## 2.23.0
 
 ## 2.22.0