+++ packages/widget.vite-plugin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osdk/widget.vite-plugin
 
+## 3.22.0
+
+### Patch Changes
+
+- @osdk/widget.api@3.22.0
+
+## 3.21.0
+
+### Patch Changes
+
+- @osdk/widget.api@3.21.0
+
 ## 3.20.0
 
 ### Patch Changes